Deleten van onbekende persoon correct geïmplementeerd + Gender als enum ipv String

+++ spring-boot/src/main/java/be/testprojects/springboot/services/PersonServiceImpl.java
@@ -1,6 +1,7 @@
 package be.testprojects.springboot.services;
 
 import be.testprojects.springboot.entities.Person;
+import be.testprojects.springboot.exceptions.RecordNotFoundException;
 import be.testprojects.springboot.repositories.PersonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,6 +23,10 @@ public class PersonServiceImpl implements PersonService {
     }
 
     public void delete(Long id) {
+        if (!personRepository.existsById(id)) {
+            throw new RecordNotFoundException();
+        }
+
         personRepository.deleteById(id);
     }
 }
